Added SSE4.1 support for faster convolutions

+++ src/main.cpp
@@ -58,7 +58,7 @@ int main(int argc, char** argv) {
 	model.deserialize(argv[1]);
 
 	// create the PartsBasedDetector and distribute the model parameters
-	PartsBasedDetector<double> pbd;
+	PartsBasedDetector<float> pbd;
 	pbd.distributeModel(model);
 
 	// load the image from file
